+++ src/utils/DownloadFFmpeg.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include <QString>
+#include <QWidget>
+
+QString installFFmpeg(QWidget* parent);