+++ src/utils/GetVideoDuration.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <QString>
+
+QString getVideoDuration(const QString &ffmpegPath, const QString &videoPath);